use methods mor easily
